WEB-1468 rename file to helpers
