+++ src/utils/helpers.tsx
@@ -1,12 +1,3 @@
-/**
- * This file implements lodash alternatives used by components
- *
- * Importing the lodash library causes problems when building Mística as an ES module.
- * Importing lodash-es causes problems when running ssr tests (not sure why)
- *
- * Once Mística gets migrated to a proper ES module, we can consider to remove this file and use lodash-es
- */
-
 type Debounced<T> = T & {cancel: () => void};
 
 export const debounce = <T extends (...args: Array<any>) => any>(
